Need to fix Histogram and NISTMottRS.

+++ CudaMONSEL/CudaMONSEL/Amphibian/Algorithm.cu
@@ -13,21 +13,21 @@ namespace Algorithm
       return l;
    }
 
-   static int partition(int a[], int l, int r)
+   static int partition(double a[], int l, int r)
    {
       int j = l - 1; // points to the last element less than pivot
       for (int i = l; i < r; ++i) { // going from l to r
          if (a[i] < a[r]) { // when sees an element less than pivot
             ++j; // points to the first element greater than or equal to pivot
             // swap i and j, so j points to last element less than pivot
-            int t = a[i];
+            double t = a[i];
             a[i] = a[j];
             a[j] = t;
          }
       }
       ++j; // points to first element greater than or equal to pivot
       // swap
-      int t = a[r];
+      double t = a[r];
       a[r] = a[j];
       a[j] = t;
 
@@ -35,7 +35,7 @@ namespace Algorithm
       return j;
    }
 
-   void quicksort(int a[], int l, int r)
+   void quicksort(double a[], int l, int r)
    {
       if (l >= r) return;
       int pv = partition(a, l, r);
